Refactor BrowserProcess into its own package
Before we can refactor to work with a registering system for the pid
that can help us to shutdown all the browser processes. We need to
abstract away the BrowserProcess into its own package to avoid
cyclic dependencies.
