+++ browserprocess/browser_process_test.go
@@ -1,4 +1,4 @@
-package common
+package browserprocess
 
 import (
 	"context"
