+++ chromium/browser_type.go
@@ -14,6 +14,7 @@ import (
 	"time"
 
 	"github.com/grafana/xk6-browser/api"
+	"github.com/grafana/xk6-browser/browserprocess"
 	"github.com/grafana/xk6-browser/common"
 	"github.com/grafana/xk6-browser/k6ext"
 	"github.com/grafana/xk6-browser/log"
@@ -219,7 +220,7 @@ func (b *BrowserType) allocate(
 	ctx context.Context, opts *common.LaunchOptions,
 	flags map[string]any, env []string, dataDir *storage.Dir,
 	logger *log.Logger,
-) (_ *common.BrowserProcess, rerr error) {
+) (_ *browserprocess.BrowserProcess, rerr error) {
 	bProcCtx, bProcCtxCancel := context.WithTimeout(ctx, opts.Timeout)
 	defer func() {
 		if rerr != nil {
@@ -237,7 +238,7 @@ func (b *BrowserType) allocate(
 		path = b.ExecutablePath()
 	}
 
-	return common.NewBrowserProcess(bProcCtx, path, args, env, dataDir, bProcCtxCancel, logger) //nolint: wrapcheck
+	return browserprocess.NewBrowserProcess(bProcCtx, path, args, env, dataDir, bProcCtxCancel, logger) //nolint: wrapcheck
 }
 
 // parseArgs parses command-line arguments and returns them.
